feat: implement useCharacterStats hook for managing character statistics
- Added a custom hook `useCharacterStats` to encapsulate logic for character stats management.
- Implemented functions for stat validation, total stat calculation, and handling stat changes.
- Included random stat generation based on character class and job selection.
- Added functionality to handle level changes and adjust stats accordingly.
- Integrated with React Hook Form for form state management and validation.
